+++ data-persistence/variables.tf
@@ -3,7 +3,7 @@ variable "DEPLOY_NAME" {
 }
 
 variable "MATURITY" {
-  type = string
+  type    = string
   default = "dev"
 }